fix show comments and fix routes for comments

+++ app/Http/Controllers/CommentController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Comment;
+use App\Http\Resources\Post\CommentsCollection;
 use App\Http\Resources\Post\PostResource;
 use App\Post;
 use Illuminate\Http\Request;
@@ -11,44 +12,64 @@ use Tymon\JWTAuth\Facades\JWTAuth;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('jwt.auth');
+
+    }
+
+//    view comments
+    public function index($id)
+    {
+        $post = Post::query()->findOrFail($id);
+        $comments = Comment::query()->where('post_id', $post->id)->get();
+        return CommentsCollection::collection($comments);
+    }
+
     // add Commment
-    public function addComment(Request $request , $id ){
-        if ($request->isMethod('post')){
+    public function addComment(Request $request, $id)
+    {
+        if ($request->isMethod('post')) {
             $this->validate($request, [
                 'comment_body' => 'required',
             ]);
             $comment = new Comment();
             $comment->comment_body = $request->input('comment_body');
-            $comment->created_at =now();
-            $comment->post_id = $id ;
-            $comment->user_id = JWTAuth::parseToken()->authenticate()->id ;
+            $comment->created_at = now();
+            $comment->post_id = $id;
+            $comment->user_id = JWTAuth::parseToken()->authenticate()->id;
             $comment->save();
+            return response()->json(['success' => 'Comment Added Successfully']);
+        } else {
+            return response()->json(['error' => 'Un Authenticated']);
         }
-        $post = Post::findOrFail($id);
-        return new PostResource($post);
     }
 
     //update comment
-    public function updateComment(Request $request,$id){
-        if ($request->isMethod('post')) {
+    public function updateComment(Request $request, $id)
+    {
+        $comment = Comment::findOrFail($id);
+        if ($request->isMethod('PUT') && $comment->user_id == JWTAuth::parseToken()->authenticate()->id) {
             $this->validate($request, [
                 'comment_body' => 'required',
             ]);
-            $comment = Comment::findOrFail($id);
-            $pid = $comment->post_id;
             $comment->comment_body = $request->input('comment_body');
             $comment->update();
-            $post = Post::findOrFail($pid);
-            return new PostResource($post);
+            return response()->json(['success' => 'Comment Updated Successfully']);
+        } else {
+            return response()->json(['error' => 'Un Authenticated']);
         }
     }
 
     //delete comment
-    public function destroyComment($id){
+    public function destroyComment($id)
+    {
         $comment = Comment::findOrFail($id);
-        $pid = $comment->post_id;
-        $comment->delete();
-        $post = Post::findOrFail($pid);
-        return new PostResource($post);
+        if ($comment->user_id == JWTAuth::parseToken()->authenticate()->id) {
+            $comment->delete();
+            return response()->json(['success' => 'Comment Deleted Successfully']);
+        } else {
+            response()->json(['error' => 'Un Authenticated']);
+        }
     }
 }
